[Chore] 파일 나누기 (#38)
